moongate v0.1.0 Moongate.Pools.Pool

Functions

handle_call(msg, from, state)

Synchronously call a function defined on the pool module, replying with the result.

handle_cast(msg, state)

Handle a pool member mutation event.

handle_info(arg1, state)

Given a callback and a number of milliseconds, cascade that callback as it is defined on the pool module on a timer set to the interval provided.

start_link(arg1)

touch_test(arg1)

world_name()