moongate v0.1.0 Moongate.Say

Functions

greeting()

A greeting message, output when the server is started.

origin(o)

Given a Moongate.SocketOrigin, return the appropriate string to use to represent that origin.

pretty(string, modifier)

pretty(string, modifier, options)

Format and output a colorized ANSI string.